feat(useIMask): add new hook for imask
