+++ src/components/DatePicker/DatePickerFieldTypeTime/helpers.ts
@@ -5,13 +5,16 @@ import {
   addSeconds,
   format,
   isValid,
+  isWithinInterval,
   parse,
   startOfDay,
   startOfHour,
   startOfMinute,
 } from 'date-fns';
 import IMask from 'imask';
-import React, { useCallback, useEffect, useRef } from 'react';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
+
+import { useIMask } from '##/components/TextField';
 
 import { useMutableRef } from '../../../hooks/useMutableRef/useMutableRef';
 import { PropsWithHTMLAttributes } from '../../../utils/types/PropsWithHTMLAttributes';
@@ -29,6 +32,7 @@ import {
 } from '../../TextField/TextField';
 import {
   datePickerPropFormatTypeTime,
+  getParts,
   getPartsDate,
   getTimeEnum,
 } from '../helpers';
@@ -82,147 +86,221 @@ export type DatePickerFieldTypeTimeProps = PropsWithHTMLAttributes<
   HTMLDivElement
 >;
 
-export const useImask = (
-  formatProp: string,
-  separator: string,
-  multiplicityHours: number | undefined,
-  multiplicitySeconds: number | undefined,
-  multiplicityMinutes: number | undefined,
-  inputRef: React.RefObject<HTMLInputElement>,
-  stringValue: string | null,
-  onError: DatePickerPropOnError | undefined,
-  handleChanhe: (props: { e: Event; value: string | null }) => void,
-) => {
-  const imaskRef = useRef<IMask.InputMask<IMask.MaskedDateOptions> | null>(
-    null,
-  );
+type UsePickerProps = {
+  value?: Date | null;
+  onChange?: DatePickerFieldTypeTimePropOnChange;
+  onError?: DatePickerPropOnError;
+  multiplicityHours: number | undefined;
+  multiplicitySeconds: number | undefined;
+  multiplicityMinutes: number | undefined;
+  format: string;
+  separator: string;
+  minDate: Date;
+  maxDate: Date;
+};
+
+export const usePicker = (props: UsePickerProps) => {
+  const {
+    value,
+    onChange,
+    onError,
+    multiplicityHours,
+    multiplicityMinutes,
+    multiplicitySeconds,
+    format: formatProp,
+    separator,
+    maxDate,
+    minDate,
+  } = props;
+  const onChangeRef = useMutableRef(onChange);
+  const valueRef = useMutableRef(value);
   const onErrorRef = useMutableRef(onError);
-  const handleChanheRef = useMutableRef(handleChanhe);
 
-  // задаем маску и сохраняем обьект маски в ref
-  // обнавляем при смене формата
-  useEffect(() => {
-    if (!inputRef.current) {
-      return;
-    }
+  const [stringValue, setStringValue] = useState<string | null>(
+    value && isValid(value) ? format(value, formatProp) : null,
+  );
+  const stringValueRef = useMutableRef(stringValue);
 
-    imaskRef.current = IMask(inputRef.current, {
-      mask: Date,
-      pattern: formatProp,
-      blocks: {
-        HH:
-          multiplicityHours && multiplicityHours > 1
-            ? {
-                mask: IMask.MaskedEnum,
-                enum: getTimeEnum(
-                  24,
-                  multiplicityHours,
-                  startOfDay,
-                  addHours,
-                  getLabelHours,
-                ),
-              }
-            : {
-                mask: IMask.MaskedRange,
-                from: 0,
-                to: 23,
-              },
-        mm:
-          multiplicityMinutes && multiplicityMinutes > 1
-            ? {
-                mask: IMask.MaskedEnum,
-                enum: getTimeEnum(
-                  60,
-                  multiplicityMinutes,
-                  startOfHour,
-                  addMinutes,
-                  getLabelMinutes,
-                ),
-              }
-            : {
-                mask: IMask.MaskedRange,
-                from: 0,
-                to: 59,
-              },
-        ss:
-          multiplicitySeconds && multiplicitySeconds > 1
-            ? {
-                mask: IMask.MaskedEnum,
-                enum: getTimeEnum(
-                  60,
-                  multiplicitySeconds,
-                  startOfMinute,
-                  addSeconds,
-                  getLabelSeconds,
-                ),
-              }
-            : {
-                mask: IMask.MaskedRange,
-                from: 0,
-                to: 59,
-              },
-      },
-      lazy: true,
-      autofix: true,
-      format: (date) => format(date, formatProp),
-      parse: (string) => parse(string, formatProp, new Date()),
-      validate: (string: string) => {
-        const [HH, mm, ss] = getPartsDate(string, formatProp, ':', false, [
+  const formatParts = useMemo(
+    () => getParts(formatProp, ':'),
+    [formatProp, separator],
+  );
+
+  const handleChange = useCallback(
+    ({ e, value: stringValue }: { e: Event; value: string | null }) => {
+      if (stringValueRef.current === stringValue) {
+        return;
+      }
+      setStringValue(stringValue);
+      const onChange = onChangeRef.current;
+      const value = valueRef.current;
+      if (onChange) {
+        if (!stringValue) {
+          if (value) {
+            onChange({ e, value: null });
+          }
+          return;
+        }
+        const partsTime = getPartsDate(stringValue, formatProp, ':', false, [
           'HH',
           'mm',
           'ss',
         ]);
 
-        if (
-          HH &&
-          mm &&
-          ss &&
-          !isValid(
-            parse(
-              `${HH}:${mm}:${ss}`,
-              datePickerPropFormatTypeTime,
-              new Date(),
-            ),
-          )
-        ) {
-          onErrorRef.current?.({
-            type: datePickerErrorTypes[1],
-            stringValue: string,
-            HH,
-            mm,
-            ss,
-          });
+        const [HH, mm, ss] = partsTime;
+        if (partsTime.filter((item) => !!item).length === formatParts.length) {
+          const date = parse(
+            `${HH}:${mm}:${ss}`,
+            datePickerPropFormatTypeTime,
+            value || new Date(),
+          );
+          if (!isWithinInterval(date, { start: minDate, end: maxDate })) {
+            onErrorRef.current?.({
+              type: datePickerErrorTypes[0],
+              stringValue,
+              date,
+              HH,
+              mm,
+              ss,
+            });
 
-          return false;
+            if (value) {
+              onChange({ e, value: null });
+            }
+            return;
+          }
+          onChange({ e, value: date });
+        } else if (value) {
+          onChange({ e, value: null });
         }
+      }
+    },
+    [minDate?.getTime(), maxDate?.getTime(), formatProp, separator],
+  );
 
-        return true;
-      },
-      // проблема в типах IMask
-    }) as unknown as IMask.InputMask<IMask.MaskedDateOptions>;
-  }, [
-    formatProp,
-    separator,
-    multiplicityHours,
-    multiplicitySeconds,
-    multiplicityMinutes,
-  ]);
+  const options: IMask.InputMask<IMask.MaskedDateOptions> = useMemo(
+    () =>
+      ({
+        mask: Date,
+        pattern: formatProp,
+        blocks: {
+          HH:
+            multiplicityHours && multiplicityHours > 1
+              ? {
+                  mask: IMask.MaskedEnum,
+                  enum: getTimeEnum(
+                    24,
+                    multiplicityHours,
+                    startOfDay,
+                    addHours,
+                    getLabelHours,
+                  ),
+                }
+              : {
+                  mask: IMask.MaskedRange,
+                  from: 0,
+                  to: 23,
+                },
+          mm:
+            multiplicityMinutes && multiplicityMinutes > 1
+              ? {
+                  mask: IMask.MaskedEnum,
+                  enum: getTimeEnum(
+                    60,
+                    multiplicityMinutes,
+                    startOfHour,
+                    addMinutes,
+                    getLabelMinutes,
+                  ),
+                }
+              : {
+                  mask: IMask.MaskedRange,
+                  from: 0,
+                  to: 59,
+                },
+          ss:
+            multiplicitySeconds && multiplicitySeconds > 1
+              ? {
+                  mask: IMask.MaskedEnum,
+                  enum: getTimeEnum(
+                    60,
+                    multiplicitySeconds,
+                    startOfMinute,
+                    addSeconds,
+                    getLabelSeconds,
+                  ),
+                }
+              : {
+                  mask: IMask.MaskedRange,
+                  from: 0,
+                  to: 59,
+                },
+        },
+        lazy: true,
+        autofix: true,
+        format: (date: Date) => format(date, formatProp),
+        parse: (string: string) => parse(string, formatProp, new Date()),
+        validate: (string: string) => {
+          const [HH, mm, ss] = getPartsDate(string, formatProp, ':', false, [
+            'HH',
+            'mm',
+            'ss',
+          ]);
+          if (
+            HH &&
+            mm &&
+            ss &&
+            !isValid(
+              parse(
+                `${HH}:${mm}:${ss}`,
+                datePickerPropFormatTypeTime,
+                new Date(),
+              ),
+            )
+          ) {
+            onErrorRef.current?.({
+              type: datePickerErrorTypes[1],
+              stringValue: string,
+              HH,
+              mm,
+              ss,
+            });
+            return false;
+          }
+          return true;
+        },
+        // проблема в типах IMask
+      } as unknown as IMask.InputMask<IMask.MaskedDateOptions>),
+    [
+      formatProp,
+      separator,
+      multiplicityHours,
+      multiplicityMinutes,
+      multiplicitySeconds,
+    ],
+  );
 
-  // Нужно для синхранизации value c Imask,
-  // так как value мы можем задать через пропс без самого ввода,
-  // и Imask требует ручной синхронихации в этом случае
-  const onAcept = useCallback((e: Event) => {
-    handleChanheRef.current({ e, value: imaskRef.current?.value || null });
-  }, []);
+  const { inputRef } = useIMask({
+    value: stringValue,
+    onChange: handleChange,
+    maskOptions: options,
+  });
 
+  // при изменении value, нужно обновить stringValue
   useEffect(() => {
-    imaskRef.current?.on('accept', onAcept);
-    return () => {
-      imaskRef.current?.off('accept', onAcept);
-    };
-  }, []);
+    if (value && isValid(value)) {
+      setStringValue(format(value, formatProp));
+    } else if (stringValue?.length === formatProp.length) {
+      // если количество введенных символов меньше чем в формате маски
+      // то не нужно мешать вводу с клавиатуры
+      // если дата была введена полностью и value пришел null,
+      // то можно считать что поле нуждается в очистке
+      setStringValue('');
+    }
+  }, [value?.getTime()]);
 
-  useEffect(() => {
-    imaskRef.current?.updateValue();
-  }, [stringValue]);
+  return {
+    stringValue,
+    inputRef,
+  };
 };
