+++ src/components/DatePicker/DatePickerFieldTypeYear/helpers.ts
@@ -1,7 +1,10 @@
 import { IconComponent, IconPropSize } from '@consta/icons/Icon';
-import { format, isValid, parse } from 'date-fns';
+import { format, isValid, isWithinInterval, parse } from 'date-fns';
 import IMask from 'imask';
-import { useEffect, useRef } from 'react';
+import { useCallback, useEffect, useMemo, useState } from 'react';
+
+import { useIMask } from '##/components/TextField';
+import { useMutableRef } from '##/hooks/useMutableRef';
 
 import { PropsWithHTMLAttributes } from '../../../utils/types/PropsWithHTMLAttributes';
 import {
@@ -59,22 +62,89 @@ export type DatePickerFieldTypeYearProps = PropsWithHTMLAttributes<
   HTMLDivElement
 >;
 
-export const useImask = (
-  formatProp: string,
-  separator: string,
-  inputRef: React.RefObject<HTMLInputElement>,
-  stringValue: string | null,
-  onError: DatePickerPropOnError | undefined,
-) => {
-  const imaskRef = useRef<IMask.InputMask<IMask.MaskedDateOptions> | null>(
-    null,
+type UsePickerProps = {
+  value?: Date | null;
+  onChange?: DatePickerFieldTypeYearPropOnChange;
+  onError?: DatePickerPropOnError;
+  format: string;
+  separator: string;
+  minDate: Date;
+  maxDate: Date;
+};
+
+export const usePicker = (props: UsePickerProps) => {
+  const {
+    value,
+    onChange,
+    onError,
+    format: formatProp,
+    separator,
+    maxDate,
+    minDate,
+  } = props;
+  const onChangeRef = useMutableRef(onChange);
+  const valueRef = useMutableRef(value);
+  const onErrorRef = useMutableRef(onError);
+
+  const [stringValue, setStringValue] = useState<string | null>(
+    value && isValid(value) ? format(value, formatProp) : null,
+  );
+  const stringValueRef = useMutableRef(stringValue);
+
+  const handleChange = useCallback(
+    ({ e, value: stringValue }: { e: Event; value: string | null }) => {
+      if (stringValueRef.current === stringValue) {
+        return;
+      }
+
+      setStringValue(stringValue);
+
+      const onChange = onChangeRef.current;
+      const value = valueRef.current;
+
+      if (onChange) {
+        if (!stringValue) {
+          if (value) {
+            onChange({ e, value: null });
+          }
+          return;
+        }
+
+        const [yyyy] = getPartsDate(stringValue, formatProp, separator, false, [
+          'yyyy',
+        ]);
+
+        if (yyyy) {
+          const date = parse(
+            `${yyyy}`,
+            datePickerPropFormatTypeYear,
+            new Date(),
+          );
+          if (!isWithinInterval(date, { start: minDate, end: maxDate })) {
+            onErrorRef.current?.({
+              type: datePickerErrorTypes[0],
+              stringValue,
+              yyyy,
+              date,
+            });
+
+            if (value) {
+              onChange({ e, value: null });
+            }
+            return;
+          }
+          onChange({ e, value: date });
+        } else if (value) {
+          onChange({ e, value: null });
+        }
+      }
+    },
+    [minDate?.getTime(), maxDate?.getTime(), formatProp, separator],
   );
 
-  // задаем маску и сохраняем обьект маски в ref
-  // обнавляем при смене формата
-  useEffect(() => {
-    if (inputRef.current) {
-      imaskRef.current = IMask(inputRef.current, {
+  const options: IMask.InputMask<IMask.MaskedDateOptions> = useMemo(
+    () =>
+      ({
         mask: Date,
         pattern: formatProp,
         blocks: {
@@ -86,8 +156,8 @@ export const useImask = (
         },
         lazy: true,
         autofix: true,
-        format: (date) => format(date, formatProp),
-        parse: (string) => parse(string, formatProp, new Date()),
+        format: (date: Date) => format(date, formatProp),
+        parse: (string: string) => parse(string, formatProp, new Date()),
         validate: (string: string) => {
           const [yyyy] = getPartsDate(string, formatProp, separator, false, [
             'yyyy',
@@ -110,14 +180,31 @@ export const useImask = (
           return true;
         },
         // проблема в типах IMask
-      }) as unknown as IMask.InputMask<IMask.MaskedDateOptions>;
-    }
-  }, [formatProp, separator]);
+      } as unknown as IMask.InputMask<IMask.MaskedDateOptions>),
+    [formatProp, separator],
+  );
 
-  // Нужно для синхранизации value c Imask,
-  // так как value мы можем задать через пропс без самого ввода,
-  // и Imask требует ручной синхронихации в этом случае
+  const { inputRef } = useIMask({
+    value: stringValue,
+    onChange: handleChange,
+    maskOptions: options,
+  });
+
+  // при изменении value, нужно обновить stringValue
   useEffect(() => {
-    imaskRef.current?.updateValue();
-  }, [stringValue]);
+    if (value && isValid(value)) {
+      setStringValue(format(value, formatProp));
+    } else if (stringValue?.length === formatProp.length) {
+      // если количество введенных символов меньше чем в формате маски
+      // то не нужно мешать вводу с клавиатуры
+      // если дата была введена полностью и value пришел null,
+      // то можно считать что поле нуждается в очистке
+      setStringValue('');
+    }
+  }, [value?.getTime()]);
+
+  return {
+    stringValue,
+    inputRef,
+  };
 };
