+++ src/components/TextField/__stand__/TextField.dev.stand.mdx
@@ -459,41 +459,29 @@ export const TextFieldExampleWtihClearButton = () => {
 
 ```tsx
 import IMask from 'imask';
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useState } from 'react';
 
-import { TextField } from '@consta/uikit/TextField';
+import { TextField, useIMask } from '@consta/uikit/TextField';
 
 export const TextFieldExampleMaskAmount = () => {
   const [value, setValue] = useState<string | null>(null);
-  const inputRef = useRef<HTMLInputElement>(null);
-  const imaskRef = useRef<IMask.InputMask<IMask.MaskedNumberOptions> | null>(
-    null,
-  );
-  useEffect(() => {
-    if (inputRef.current) {
-      imaskRef.current = IMask(inputRef.current, {
-        mask: Number,
-        min: 0,
-        scale: 2,
-        thousandsSeparator: ' ',
-        radix: ',',
-        padFractionalZeros: true,
-        normalizeZeros: true,
-        mapToRadix: ['.'],
-      });
-    }
-  }, []);
 
-  useEffect(() => {
-    imaskRef.current?.on('accept', () =>
-      setValue(imaskRef.current?.value || null),
-    );
-    return () => {
-      imaskRef.current?.off('accept', () =>
-        setValue(imaskRef.current?.value || null),
-      );
-    };
-  }, []);
+  const { inputRef } = useIMask({
+    value,
+    onChange: ({ value }) => setValue(value),
+    maskOptions: {
+      mask: Number,
+      min: 0,
+      scale: 2,
+      thousandsSeparator: ' ',
+      radix: ',',
+      padFractionalZeros: true,
+      autofix: true,
+      lazy: true,
+      normalizeZeros: true,
+      mapToRadix: ['.'],
+    },
+  });
 
   return (
     <TextField
@@ -512,35 +500,18 @@ export const TextFieldExampleMaskAmount = () => {
 
 ```tsx
 import IMask from 'imask';
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useState } from 'react';
 
-import { TextField } from '@consta/uikit/TextField';
+import { TextField, useIMask } from '@consta/uikit/TextField';
 
 export const TextFieldExampleMaskPhone = () => {
   const [value, setValue] = useState<string | null>(null);
-  const inputRef = useRef<HTMLInputElement>(null);
-  const mask = '+{7}(000)000-00-00';
-  const imaskRef = useRef<IMask.InputMask<IMask.MaskedPatternOptions> | null>(
-    null,
-  );
-  useEffect(() => {
-    if (inputRef.current) {
-      imaskRef.current = IMask(inputRef.current, {
-        mask,
-      });
-    }
-  }, []);
 
-  useEffect(() => {
-    imaskRef.current?.on('accept', () =>
-      setValue(imaskRef.current?.value || null),
-    );
-    return () => {
-      imaskRef.current?.off('accept', () =>
-        setValue(imaskRef.current?.value || null),
-      );
-    };
-  }, []);
+  const { inputRef } = useIMask({
+    value,
+    onChange: ({ value }) => setValue(value),
+    maskOptions: '+{7}(000)000-00-00',
+  });
 
   return (
     <TextField
@@ -559,35 +530,18 @@ export const TextFieldExampleMaskPhone = () => {
 
 ```tsx
 import IMask from 'imask';
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useState } from 'react';
 
-import { TextField } from '@consta/uikit/TextField';
+import { TextField, useIMask } from '@consta/uikit/TextField';
 
 export const TextFieldExampleMaskPassport = () => {
   const [value, setValue] = useState<string | null>(null);
-  const inputRef = useRef<HTMLInputElement>(null);
-  const mask = '00 00 000000';
-  const imaskRef = useRef<IMask.InputMask<IMask.MaskedPatternOptions> | null>(
-    null,
-  );
-  useEffect(() => {
-    if (inputRef.current) {
-      imaskRef.current = IMask(inputRef.current, {
-        mask,
-      });
-    }
-  }, []);
 
-  useEffect(() => {
-    imaskRef.current?.on('accept', () =>
-      setValue(imaskRef.current?.value || null),
-    );
-    return () => {
-      imaskRef.current?.off('accept', () =>
-        setValue(imaskRef.current?.value || null),
-      );
-    };
-  }, []);
+  const { inputRef } = useIMask({
+    value,
+    onChange: ({ value }) => setValue(value),
+    maskOptions: '00 00 000000',
+  });
 
   return (
     <TextField
@@ -606,35 +560,18 @@ export const TextFieldExampleMaskPassport = () => {
 
 ```tsx
 import IMask from 'imask';
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useState } from 'react';
 
-import { TextField } from '@consta/uikit/TextField';
+import { TextField, useIMask } from '@consta/uikit/TextField';
 
 export const TextFieldExampleMaskBankAccount = () => {
   const [value, setValue] = useState<string | null>(null);
-  const inputRef = useRef<HTMLInputElement>(null);
-  const mask = '0000 0000 0000 0000';
-  const imaskRef = useRef<IMask.InputMask<IMask.MaskedPatternOptions> | null>(
-    null,
-  );
-  useEffect(() => {
-    if (inputRef.current) {
-      imaskRef.current = IMask(inputRef.current, {
-        mask,
-      });
-    }
-  }, []);
 
-  useEffect(() => {
-    imaskRef.current?.on('accept', () =>
-      setValue(imaskRef.current?.value || null),
-    );
-    return () => {
-      imaskRef.current?.off('accept', () =>
-        setValue(imaskRef.current?.value || null),
-      );
-    };
-  }, []);
+  const { inputRef } = useIMask({
+    value,
+    onChange: ({ value }) => setValue(value),
+    maskOptions: '0000 0000 0000 0000',
+  });
 
   return (
     <TextField
@@ -652,7 +589,7 @@ export const TextFieldExampleMaskBankAccount = () => {
 ## Импорт
 
 ```tsx
-import { TextField } from '@consta/uikit/TextField';
+import { TextField, useIMask } from '@consta/uikit/TextField';
 ```
 
 ## Свойства
