+++ src/components/TextField/index.ts
@@ -1,1 +1,2 @@
 export * from './TextField';
+export * from './useIMask';
